upgrading civicrm to alpha1 on new branch

+++ drupal/sites/all/modules/civicrm/packages/dompdf/include/cpdf_adapter.cls.php
@@ -4,8 +4,10 @@
  *
  * File: $RCSfile: cpdf_adapter.cls.php,v $
  * Created on: 2004-08-04
+ * Modified on: 2008-01-05
  *
  * Copyright (c) 2004 - Benj Carson <benjcarson@digitaljunkies.ca>
+ * Portions copyright (c) 2008 - Orion Richardson <orionr@yahoo.com>
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -28,16 +30,29 @@
  * the case, you can obtain a copy at http://www.php.net/license/3_0.txt.
  *
  * The latest version of DOMPDF might be available at:
- * http://www.digitaljunkies.ca/dompdf
+ * http://www.dompdf.com/
  *
- * @link http://www.digitaljunkies.ca/dompdf
+ * @link http://www.dompdf.com/
  * @copyright 2004 Benj Carson
  * @author Benj Carson <benjcarson@digitaljunkies.ca>
+ * @contributor Orion Richardson <orionr@yahoo.com>
+ * @contributor Helmut Tischer <htischer@weihenstephan.org>
  * @package dompdf
- * @version 0.5.1
+ *
+ * Changes
+ * @contributor Helmut Tischer <htischer@weihenstephan.org>
+ * @version 0.5.1.htischer.20090507
+ * - On gif to png conversion tmp file creation, clarify tmp name and add to tmp deletion list only on success
+ * - On gif to png conversion, when available add direct from gd without tmp file, skip image load if already cached.
+ *   to safe CPU time and memory
+ * @contributor Helmut Tischer <htischer@weihenstephan.org>
+ * @version dompdf_trunk_with_helmut_mods.20090524
+ * - Pass temp and fontcache folders to Cpdf, to making Cpdf independent from dompdf
+ * @version dompdf_trunk_with_helmut_mods.20090528
+ * - fix text position according to glyph baseline to match background rectangle
  */
 
-/* $Id: cpdf_adapter.cls.php,v 1.16 2006/07/07 21:31:03 benjcarson Exp $ */
+/* $Id: cpdf_adapter.cls.php 217 2010-03-11 23:03:57Z ryan.masten $ */
 
 // FIXME: Need to sanity check inputs to this class
 require_once(DOMPDF_LIB_DIR . "/class.pdf.php");
@@ -184,34 +199,29 @@ class CPDF_Adapter implements Canvas {
    * @param mixed  $paper  The size of paper to use in this PDF ({@link CPDF_Adapter::$PAPER_SIZES})
    * @param string $orientation The orienation of the document (either 'landscape' or 'portrait')
    */
-  function __construct($paper = "letter", $orientation = "portrait") {    
+  function __construct($paper = "letter", $orientation = "portrait") {
 
     if ( is_array($paper) )
       $size = $paper;
     else if ( isset(self::$PAPER_SIZES[mb_strtolower($paper)]) )
-      $size = self::$PAPER_SIZES[$paper];
+      $size = self::$PAPER_SIZES[mb_strtolower($paper)];
     else
       $size = self::$PAPER_SIZES["letter"];
 
-    if ( mb_strtolower($orientation) == "landscape" ) {
+    if ( mb_strtolower($orientation) === "landscape" ) {
       $a = $size[3];
       $size[3] = $size[2];
       $size[2] = $a;
     }
     
-    $this->_pdf = new Cpdf($size);
-    $this->_pdf->addInfo("Creator", "DOMPDF Converter");
+    $this->_pdf = new Cpdf($size, DOMPDF_UNICODE_ENABLED, DOMPDF_FONT_CACHE, DOMPDF_TEMP_DIR);
+    $this->_pdf->addInfo("Creator", "dompdf");
 
     // Silence pedantic warnings about missing TZ settings
-    if ( function_exists("date_default_timezone_get") ) {
-      $tz = @date_default_timezone_get();
-      date_default_timezone_set("UTC");
-      $this->_pdf->addInfo("CreationDate", date("Y-m-d"));
-      date_default_timezone_set($tz);
-
-    } else {
-      $this->_pdf->addInfo("CreationDate", date("Y-m-d"));
-    }
+    $time = substr_replace(date('YmdHisO'), '\'', (0 - 2), 0).'\'';
+	$this->_pdf->addInfo("CreationDate", "D:".$time);
+	$this->_pdf->addInfo("ModDate", "D:".$time);
+    
 
     $this->_width = $size[2] - $size[0];
     $this->_height= $size[3] - $size[1];
@@ -232,7 +242,10 @@ class CPDF_Adapter implements Canvas {
    */
   function __destruct() {
     foreach ($this->_image_cache as $img) {
-      unlink($img);
+      //debugpng
+      if (DEBUGPNG) print '[__destruct unlink '.$img.']';
+      if (!DEBUGKEEPTEMP)
+        unlink($img);
     }
   }
   
@@ -242,6 +255,16 @@ class CPDF_Adapter implements Canvas {
    * @return Cpdf
    */
   function get_cpdf() { return $this->_pdf; }
+
+  /**
+   * Add meta information to the PDF
+   *
+   * @param string $label  label of the value (Creator, Producter, etc.)
+   * @param string $value  the text to set
+   */
+  function add_info($label, $value) {
+    $this->_pdf->addInfo($label, $value);
+  }
 
   /**
    * Opens a new 'object'
@@ -271,7 +294,7 @@ class CPDF_Adapter implements Canvas {
    */
   function reopen_object($object) {
     $this->_pdf->reopenObject($object);
-    $this->_pdf->saveState();    
+    $this->_pdf->saveState();
   }
 
   /**
@@ -391,7 +414,7 @@ class CPDF_Adapter implements Canvas {
    * @param array $color
    */
   protected function _set_fill_color($color) {
-    list($r, $g, $b) = $color;      
+    list($r, $g, $b) = $color;
     $this->_pdf->setColor($r, $g, $b);
   }
 
@@ -473,14 +496,12 @@ class CPDF_Adapter implements Canvas {
   /**
    * Convert a GIF image to a PNG image
    *
-   * @return string The url of the newly converted image 
+   * @return string The url of the newly converted image
    */
   protected function _convert_gif_to_png($image_url) {
-    global $_dompdf_warnings;
     
     if ( !function_exists("imagecreatefromgif") ) {
-      $_dompdf_warnings[] = "Function imagecreatefromgif() not found.  Cannot convert gif image: $image_url.";      
-      return DOMPDF_LIB_DIR . "/res/broken_image.png";
+      throw new DOMPDF_Exception("Function imagecreatefromgif() not found.  Cannot convert gif image: $image_url.  Please install the image PHP extension.");
     }
 
     $old_err = set_error_handler("record_warnings");
@@ -488,8 +509,10 @@ class CPDF_Adapter implements Canvas {
 
     if ( $im ) {
       imageinterlace($im, 0);
-    
-      $filename = tempnam(DOMPDF_TEMP_DIR, "dompdf_img_");
+
+      $filename = tempnam(DOMPDF_TEMP_DIR, "gifdompdf_img_").'.png';
+      $this->_image_cache[] = $filename;
+
       imagepng($im, $filename);
 
     } else {
@@ -499,8 +522,6 @@ class CPDF_Adapter implements Canvas {
 
     restore_error_handler();
 
-    $this->_image_cache[] = $filename;
-    
     return $filename;
     
   }
@@ -562,33 +583,67 @@ class CPDF_Adapter implements Canvas {
     if ( !$fill && isset($width) )
       $this->_set_line_style($width, "round", "round", $style);
 
-    $this->_pdf->filledEllipse($x, $this->y($y), $r1, 0, 0, 8, 0, 360, 1, $fill);
-
+    $this->_pdf->ellipse($x, $this->y($y), $r1, 0, 0, 8, 0, 360, 1, $fill);
   }
   
   //........................................................................
 
   function image($img_url, $img_type, $x, $y, $w, $h) {
+    //debugpng
+    if (DEBUGPNG) print '[image:'.$img_url.'|'.$img_type.']';
 
     $img_type = mb_strtolower($img_type);
-    
+
     switch ($img_type) {
     case "jpeg":
     case "jpg":
+      //debugpng
+      if (DEBUGPNG)  print '!!!jpg!!!';
+
       $this->_pdf->addJpegFromFile($img_url, $x, $this->y($y) - $h, $w, $h);
       break;
 
     case "png":
+      //debugpng
+      if (DEBUGPNG)  print '!!!png!!!';
+
       $this->_pdf->addPngFromFile($img_url, $x, $this->y($y) - $h, $w, $h);
       break;
 
     case "gif":
       // Convert gifs to pngs
-      $img_url = $this->_convert_gif_to_png($img_url);
-      $this->_pdf->addPngFromFile($img_url, $x, $this->y($y) - $h, $w, $h);
+      //DEBUG_IMG_TEMP
+      //if (0) {
+      if ( method_exists( $this->_pdf, "addImagePng" ) ) {
+        //debugpng
+        if (DEBUGPNG)  print '!!!gif addImagePng!!!';
+
+      	//If optimization to direct png creation from gd object is available,
+        //don't create temp file, but place gd object directly into the pdf
+	    if ( method_exists( $this->_pdf, "image_iscached" ) &&
+	         $this->_pdf->image_iscached($img_url) ) {
+	      //If same image has occured already before, no need to load because
+	      //duplicate will anyway be eliminated.
+	      $img = null;
+	    } else {
+    	  $img = @imagecreatefromgif($img_url);
+    	  if (!$img) {
+      	    return;
+    	  }
+    	  imageinterlace($img, 0);
+    	}
+    	$this->_pdf->addImagePng($img_url, $x, $this->y($y) - $h, $w, $h, $img);
+      } else {
+        //debugpng
+        if (DEBUGPNG)  print '!!!gif addPngFromFile!!!';
+        $img_url = $this->_convert_gif_to_png($img_url);
+        $this->_pdf->addPngFromFile($img_url, $x, $this->y($y) - $h, $w, $h);
+      }
       break;
-      
-    default:      
+
+    default:
+      //debugpng
+      if (DEBUGPNG) print '!!!unknown!!!';
       break;
     }
     
@@ -608,8 +663,45 @@ class CPDF_Adapter implements Canvas {
     $font .= ".afm";
     
     $this->_pdf->selectFont($font);
-    $this->_pdf->addText($x, $this->y($y) - Font_Metrics::get_font_height($font, $size), $size, utf8_decode($text), $angle, $adjust);
+    
+    //Font_Metrics::get_font_height($font, $size) ==
+    //$this->get_font_height($font, $size) ==
+    //$this->_pdf->selectFont($font),$this->_pdf->getFontHeight($size)
+    //- FontBBoxheight+FontHeightOffset, scaled to $size, in pt
+    //$this->_pdf->getFontDescender($size)
+    //- Descender scaled to size
+    //
+    //$this->_pdf->fonts[$this->_pdf->currentFont] sizes:
+    //['FontBBox'][0] left, ['FontBBox'][1] bottom, ['FontBBox'][2] right, ['FontBBox'][3] top
+    //Maximum extent of all glyphs of the font from the baseline point
+    //['Ascender'] maximum height above baseline except accents
+    //['Descender'] maximum depth below baseline, negative number means below baseline
+    //['FontHeightOffset'] manual enhancement of .afm files to trim windows fonts. currently not used.
+    //Values are in 1/1000 pt for a font size of 1 pt
+    //
+    //['FontBBox'][1] should be close to ['Descender']
+    //['FontBBox'][3] should be close to ['Ascender']+Accents
+    //in practice, FontBBox values are a little bigger
+    //
+    //The text position is referenced to the baseline, not to the lower corner of the FontBBox,
+    //for what the left,top corner is given.
+    //FontBBox spans also the background box for the text.
+    //If the lower corner would be used as reference point, the Descents of the glyphs would
+    //hang over the background box border.
+    //Therefore compensate only the extent above the Baseline.
+    //
+    //print '<pre>['.$font.','.$size.','.$this->_pdf->getFontHeight($size).','.$this->_pdf->getFontDescender($size).','.$this->_pdf->fonts[$this->_pdf->currentFont]['FontBBox'][3].','.$this->_pdf->fonts[$this->_pdf->currentFont]['FontBBox'][1].','.$this->_pdf->fonts[$this->_pdf->currentFont]['FontHeightOffset'].','.$this->_pdf->fonts[$this->_pdf->currentFont]['Ascender'].','.$this->_pdf->fonts[$this->_pdf->currentFont]['Descender'].']</pre>';
+    //
+    //$this->_pdf->addText($x, $this->y($y) - Font_Metrics::get_font_height($font, $size), $size, $text, $angle, $adjust);
+	//$this->_pdf->addText($x, $this->y($y) - $size, $size, $text, $angle, $adjust);
+	//$this->_pdf->addText($x, $this->y($y) - $this->_pdf->getFontHeight($size)-$this->_pdf->getFontDescender($size), $size, $text, $angle, $adjust);
+	$this->_pdf->addText($x, $this->y($y) - ($this->_pdf->fonts[$this->_pdf->currentFont]['FontBBox'][3]*$size)/1000, $size, $text, $angle, $adjust);
+  }
 
+  //........................................................................
+
+  function javascript($code) {
+    $this->_pdf->addJavascript($code);
   }
 
   //........................................................................
@@ -654,7 +746,10 @@ class CPDF_Adapter implements Canvas {
 
   function get_text_width($text, $font, $size, $spacing = 0) {
     $this->_pdf->selectFont($font);
-    return $this->_pdf->getTextWidth($size, utf8_decode($text), $spacing);
+    if (!DOMPDF_UNICODE_ENABLED) {
+    	$text = mb_convert_encoding($text, 'Windows-1252', 'UTF-8');
+    }
+    return $this->_pdf->getTextWidth($size, $text, $spacing);
   }
 
   //........................................................................
@@ -664,6 +759,8 @@ class CPDF_Adapter implements Canvas {
     return $this->_pdf->getFontHeight($size);
   }
 
+  //........................................................................
+  
   /**
    * Writes text at the specified x and y coordinates on every page
    *
@@ -682,9 +779,27 @@ class CPDF_Adapter implements Canvas {
    * @param float $angle angle to write the text at, measured CW starting from the x-axis
    */
   function page_text($x, $y, $text, $font, $size, $color = array(0,0,0),
-                     $adjust = 0, $angle = 0,  $blend = "Normal", $opacity = 1.0) {
+                     $adjust = 0, $angle = 0) {
+    $_t = "text";
+    $this->_page_text[] = compact("_t", "x", "y", "text", "font", "size", "color", "adjust", "angle");
+  }
+
+  //........................................................................
     
-    $this->_page_text[] = compact("x", "y", "text", "font", "size", "color", "adjust", "angle");
+  /**
+   * Processes a script on every page
+   *
+   * The variables $pdf, $PAGE_NUM, and $PAGE_COUNT are available.
+   *
+   * This function can be used to add page numbers to all pages
+   * after the first one, for example.
+   *
+   * @param string $code the script code
+   * @param string $type the language type for script
+   */
+  function page_script($code, $type = "text/php") {
+    $_t = "script";
+    $this->_page_text[] = compact("_t", "code", "type");
   }
   
   //........................................................................
@@ -708,22 +823,33 @@ class CPDF_Adapter implements Canvas {
       return;
 
     $page_number = 1;
+    $eval = null;
 
     foreach ($this->_pages as $pid) {
+      $this->reopen_object($pid);
 
       foreach ($this->_page_text as $pt) {
         extract($pt);
 
+        switch ($_t) {
+            
+        case "text":
         $text = str_replace(array("{PAGE_NUM}","{PAGE_COUNT}"),
                             array($page_number, $this->_page_count), $text);
-
-        $this->reopen_object($pid);        
         $this->text($x, $y, $text, $font, $size, $color, $adjust, $angle);
-        $this->close_object();        
+          break;
+          
+        case "script":
+          if (!$eval) {
+            $eval = new PHP_Evaluator($this);
+          }
+          $eval->evaluate($code, array('PAGE_NUM' => $page_number, 'PAGE_COUNT' => $this->_page_count));
+          break;
+        }
       }
 
+      $this->close_object();
       $page_number++;
-      
     }
   }
   
@@ -771,5 +897,3 @@ class CPDF_Adapter implements Canvas {
   function get_messages() { return $this->_pdf->messages; }
   
 }
-
-?>
